sell-arts-dj-api/pyproject.toml
[tool.poetry]
name = "sellarts"
version = "0.1.0"
description = ""
authors = ["satoshi <nathanuchia@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.1.3"
graphene-django = "^3.2.2"
graphene-django-query-optimizer = "0.10.6"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
django-graphql-jwt = "^0.3.4"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
django-allow-cidr = "^0.7.1"
uvicorn = "^0.32.0"
graphene-luna = "1.0.0"
pytest-django = "^4.9.0"
pytest = "^8.3.3"
tqdm = "^4.66.6"
rich = "^ 13.9.4"
psycopg = "^ 3.2.3"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
requests = "^2.32.3"
faker = "^35.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = {version = "^24.10.0", allow-prereleases = true}
mypy = "^1.13.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

